contactbookapi/requirements.txt
asgiref==3.5.2
Django==4.1.7
django-cors-headers==3.13.0
djangorestframework==3.14.0
python-decouple==3.6
sqlparse==0.4.3
tzdata==2022.5

# Development dependencies
isort